Move platform exports to WinRT.Runtime
